db/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.1
python-dotenv==0.21.0
sqlparse==0.4.2
djangorestframework==3.13.1
django-cors-headers==3.13.0